renamed partlets directory
