+++ src/report/partlets/__init__.py
@@ -1,0 +1,3 @@
+from src.report.partlets.partlet import Partlet
+from src.report.partlets.header import Header
+from src.report.partlets.weight import WeightDiary
